requirements.txt
slack-sdk
slack-bolt
python-dotenv
google-generativeai